refactor(balancing): clarify Template and Model boundaries
