+++ libs/gda-balancing/src/gda_balancing/domain/template/_release_semantics.py
