+++ libs/gda-balancing/tests/test_schema2_template_cli.py
@@ -924,6 +924,7 @@ def test_template_list_exposes_the_packaged_content_addressed_release(run_cli):
 
 def test_template_schema_identity_refuses_a_missing_authority_contract():
     authority = authority_set()
+    kernel = authority["kernel"]
     language_bundle = authority["language_bundle"]
     source_schema = cast(Any, language_bundle["language"])["wire_schemas"][0]
     schema_kind = source_schema["artifact_kind"]
@@ -933,7 +934,7 @@ def test_template_schema_identity_refuses_a_missing_authority_contract():
         ValueError,
         match=f"exact wire-schema identity domain is unavailable for {schema_kind}",
     ):
-        wire_schema_identity_for_kind(language_bundle, schema_kind)
+        minimal_release(kernel, language_bundle)
 
 
 def test_every_wire_schema_consumer_projects_an_extension_owned_identity_domain():
@@ -946,6 +947,15 @@ def test_every_wire_schema_consumer_projects_an_extension_owned_identity_domain(
     context = admit_authority_context(kernel, language_bundle)
 
     assert isinstance(context, AdmittedAuthorityContext)
+    release = minimal_release(
+        cast(Any, context.kernel),
+        cast(Any, context.language_bundle),
+    )
+    model_source_member = next(
+        member
+        for member in cast(list[dict[str, Any]], release["members"])
+        if member["member_kind"] == "model-source-package"
+    )
     schema_body = {
         key: value for key, value in source_schema["schema"].items() if key != "$id"
     }
@@ -969,6 +979,7 @@ def test_every_wire_schema_consumer_projects_an_extension_owned_identity_domain(
         )
         == expected
     )
+    assert model_source_member["member_schema_identity"] == expected
     assert projected_source_schema["$id"].endswith(expected.removeprefix("sha256:"))
 
 
@@ -1121,6 +1132,7 @@ def test_template_admits_a_member_whose_artifact_and_schema_kinds_differ():
 
 def test_template_refuses_an_artifact_kind_that_shadows_a_standalone_schema():
     authority = authority_set()
+    kernel = authority["kernel"]
     language_bundle = authority["language_bundle"]
     original_identity = wire_schema_identity_for_kind(
         language_bundle,
@@ -1135,7 +1147,12 @@ def test_template_refuses_an_artifact_kind_that_shadows_a_standalone_schema():
         ValueError,
         match="wire-schema kind authority is ambiguous: negative-vector",
     ):
-        wire_schema_identity_for_kind(language_bundle, "negative-vector")
+        minimal_release(kernel, language_bundle)
+    with pytest.raises(
+        ValueError,
+        match="wire-schema kind authority is ambiguous: negative-vector",
+    ):
+        artifact_wire_schema_identity(language_bundle, "negative-vector")
 
     assert original_identity == content_identity(
         "negative-vector-wire-schema-v2",
